requirements.txt
torch==2.4.0
pdfplumber==0.9.0
camelot-py==1.0.0
loguru==0.7.0
langchain==0.1.1
transformers==4.45.1
sentencepiece==0.2.0
vllm==0.6.1.post1
peft==0.14.0
scikit-learn==1.6.1
trl==0.9.4
gradio==4.18.0
galore_torch==1.0
opencv-python==4.11.0.86
parse==1.20.2
fastbm25==0.0.2
ghostscript==0.7
datasets==3.2.0
jieba==0.42.1
rouge-chinese==1.0.3
nltk==3.9.1
sse_starlette==2.2.1
logger==1.4
text2vec==1.3.3